refactor(web): share model OAuth sign-in lifecycle and desktop callback type
Use RakazoDesktopOAuthCallback from contracts in the loopback parser, and move
the duplicated capture/cancel/submit ownership logic into useModelOAuthSignIn
so onboarding and model settings keep one attempt-safe path.
